nihal

+++ app/src/main/java/com/example/appdedestrubition/products.kt
@@ -1,3 +1,3 @@
 package com.example.appdedestrubition
 
-data class products(var images:Int ,var nom_produit: String ,var prix : Int,var total:Int  )
+data class products(var images:Int ,var nom_produit: String ,var prix : Double,var total:Double  )
